Move manifest into util header
