+++ valhalla/mjolnir/graphbuilder.h
@@ -1,17 +1,13 @@
 #ifndef VALHALLA_MJOLNIR_GRAPHBUILDER_H
 #define VALHALLA_MJOLNIR_GRAPHBUILDER_H
 
+#include <boost/property_tree/ptree.hpp>
 #include <cstdint>
-#include <fstream>
 #include <string>
 #include <vector>
 
-#include <boost/property_tree/ptree.hpp>
-
-#include <valhalla/baldr/graphid.h>
-#include <valhalla/baldr/rapidjson_utils.h>
 #include <valhalla/baldr/signinfo.h>
-#include <valhalla/midgard/logging.h>
+
 #include <valhalla/mjolnir/osmdata.h>
 #include <valhalla/mjolnir/osmnode.h>
 #include <valhalla/mjolnir/osmway.h>
@@ -40,7 +36,7 @@ public:
    * @param  complex_to_restriction_file    where to store the to complex restrictions so they are not
    * in memory
    */
-  static void Build(const ptree& pt,
+  static void Build(const boost::property_tree::ptree& pt,
                     const OSMData& osmdata,
                     const std::string& ways_file,
                     const std::string& way_nodes_file,
@@ -68,72 +64,6 @@ public:
                                  bool ramp,
                                  bool tc);
 };
-
-// The tile manifest is a JSON-serializable index of tiles to be processed during the build stage of
-// valhalla_build_tiles'. It can be used to distribute shard keys when building tiles with
-// parallelized, distributed batch processing. For example, a workflow orchestrator can partition
-// and distribute this manifest to workers in a 'build' stage worker pool.
-//
-// This file written out during 'constructedges' and a prerequisite for the 'build' stage run by
-// valhalla_build_tiles.
-//
-// Example manifest :
-//
-// {
-//   "tiles": [
-//     {
-//       "node_index": 0,
-//       "graphid": {
-//         "value": 5970538,
-//         "id": 0,
-//         "tile_id": 746317,
-//         "level": 2
-//       }
-//     }
-//   ]
-// }
-struct TileManifest {
-
-  std::map<baldr::GraphId, size_t> tileset;
-
-  std::string ToString() const {
-    baldr::json::ArrayPtr array = baldr::json::array({});
-    for (const auto& tile : tileset) {
-      const baldr::json::Value& graphid = tile.first.json();
-      const baldr::json::MapPtr& item = baldr::json::map(
-          {{"graphid", graphid}, {"node_index", static_cast<uint64_t>(tile.second)}});
-      array->emplace_back(item);
-    }
-    std::stringstream manifest;
-    manifest << *baldr::json::map({{"tiles", array}});
-    return manifest.str();
-  }
-
-  void LogToFile(const std::string& filename) const {
-    std::ofstream handle;
-    handle.open(filename);
-    handle << ToString();
-    handle.close();
-    LOG_INFO("Writing tile manifest to " + filename);
-  }
-
-  static TileManifest ReadFromFile(const std::string& filename) {
-    ptree manifest;
-    rapidjson::read_json(filename, manifest);
-    LOG_INFO("Reading tile manifest from " + filename);
-    std::map<baldr::GraphId, size_t> tileset;
-    for (const auto& tile_info : manifest.get_child("tiles")) {
-      const ptree& graph_id = tile_info.second.get_child("graphid");
-      const baldr::GraphId id(graph_id.get<uint64_t>("value"));
-      const size_t node_index = tile_info.second.get<size_t>("node_index");
-      tileset.insert({id, node_index});
-    }
-    LOG_INFO("Reading " + std::to_string(tileset.size()) + " tiles from tile manifest file " +
-             filename);
-    return TileManifest{tileset};
-  }
-}; // namespace mjolnir
-
 } // namespace mjolnir
 } // namespace valhalla
